added README.md, refactored package name without snake_case
