--- main.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"fmt"
-
-	"problems/internal/math"
-)
-
-func main() {
-	fmt.Println(math.IsPrime(17))
-}
