+++ task/divide_string/pkg/dividestring/dividestring_test.go
@@ -1,4 +1,4 @@
-package divide_string
+package dividestring
 
 import (
 	"fmt"
